Split component file, send (report member support request) to api
- also added empty test files for each dummy subcomponent
